Dockerfile:  Pin `linux-libc-dev` package; others
The Machinekit package repo supplies the `linux-libc-dev` package in a
version different from upstream and only in x86 architectures.  That
breaks `Multi-Arch:` because both x86 ard arm versions share files and
must be installed simultaneously.

Fix that situation by pinning the package to upstream versions to
guarantee a match.

Deal with the force-installed packages by downloading them at build
time, and supplying a script to add/remove them on demand.  That way
apt isn't in a broken configuration out of the box.

Also tweak some commands and comments.